QR code Generator/requirements.txt
pyqrcode ==1.2.1
pypng== 0.20220715.0